US Constitution Loading Screens
Part of the educational aspect of the project was to intersperse a highly fictionalized legal drama narrative with resources that would make the work more accessible and hint at the kind of studies required to practice law. Hence the beginning of the narrative explicitly outlines the adversarial system of common law practice. These image files depicted U.S. Constitutional documents were then intended to appear during loading screens to introduce legal precedents (with the intention that laws and court cases might also be displayed in the same manner, as the project expanded).